requirements.txt
fastapi[all]
uvicorn[standard]
sqlalchemy
pymysql
apscheduler